Add fully working autorandr config
The nixos module does not support all the attributes needed for
`autorandr --cycle` to work. This is remedied by manaully copying
configs generated by `autorandr` to the relevant location

Also sets the internal display as `default` so that hotplugging and
wake from sleep switching works. Counterintuitively `default` won't
actually be selected when hotplugging - the "first detected profile" ie
`autorandr -c` will be executed instead, `default` is a fallback for
when no profiles are detecte